update structs and events - use uint32 instead of uint24 for reward epoch ids

+++ contracts/userInterfaces/tee/ITeeInstructions.sol
@@ -10,7 +10,7 @@ interface ITeeInstructions {
 
     event TeeInstructionsSent(
         bytes32 indexed instructionId,
-        uint24 indexed rewardEpochId,
+        uint32 indexed rewardEpochId,
         ITeeRegistry.TeeMachine[] teeMachines,
         bytes32 opType,
         bytes32 opCommand,
